+++ src/Network/Network/Models/AzureFirewallPolicy/PSAzureFirewallPolicyDnsSetting.cs
@@ -13,26 +13,16 @@
 // limitations under the License.
 //
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Commands.Network.Models
 {
-    public class PSAzureFirewallPolicy : PSTopLevelResource
+    public class PSAzureFirewallPolicyDnsSetting
     {
-        public string ThreatIntelMode { get; set; }
+        public IList<string> Server { get; set; }
 
-        public PSAzureFirewallPolicyThreatIntelWhitelist ThreatIntelWhitelist { get; set; }
+        public bool? EnableProxy { get; set; }
 
-        public Microsoft.Azure.Management.Network.Models.SubResource BasePolicy { get; set; }
-
-        public string ProvisioningState { get; set; }
-
-        [JsonProperty("ruleCollectionGroups")]
-        public List<Microsoft.Azure.Management.Network.Models.SubResource> RuleCollectionGroups { get; set; }
-
-        public PSAzureFirewallPolicyDnsSetting DnsSetting { get; set; }
+        public bool? RequireProxyForNetworkRules { get; set; }
     }
 }